Create del
I present to you my tool developed in shellscript to move files and directories from the command line to the recycle bin of the graphical interface of the system.

An evolution of the 'rm' command that permanently deletes files and directories, the del command gives the user the opportunity to restore their files and directories from the system recycle bin, in addition to many other features such as recursive deletion: 

Do you know when you delete a file and still get residues of it in the system? the recursive deletion (del -R) erases up to 'other generations' of the file! XD